pyproject.toml
[tool.poetry]
name = "voice-assistant"
version = "0.1.0"
description = ""
authors = ["razanalshbaha <razanalshabah19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.1"
langchain-openai = "^0.1.7"
fastapi = "^0.111.0"
langchain-community = "^0.2.1"
google-api-python-client = "^2.131.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
google = "^3.0.0"
azure-cognitiveservices-speech = "^1.37.0"
pyowm = "^3.3.0"
streamlit = "^1.35.0"
langchainhub = "^0.1.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
